ACL implementation corrected as far as possible, but ...
... Spring security issue #5508 (PSQLException) will not be solved
before 5.2.1, we are using 5.1.5; and above that from one comment:
 rwinch commented on 28 Nov 2018

@nenaraab There isn't a sample with Postgres. The contact sample is the
reference for ACL support. However, I would caution you that we don't
typically recommend using ACL support because it requires doing in
memory joins. If you have a million records and the user only is able to
access 2, then you must process all the records in memory.

Instead, we recommend using the Spring Data support.

Reverting to only use users, groups and authorities and dropping use of
ACLs.